:recycle: refactor subroutines in qsyn_main.cpp to another file
